+++ src/editors/Flow/view/Components/FlowTopBar/FlowSearch.jsx
@@ -9,13 +9,28 @@ import {
 } from "@material-ui/core";
 import { Autocomplete } from "@material-ui/lab";
 import { flowTopBarStyles } from "./styles";
-import { NodeInstance } from "../../../model/subModels";
 import { useTranslation } from "react-i18next";
 
 const FlowSearch = props => {
   const { options, onChange, onEnabled, onDisabled, visible } = props;
   const classes = flowTopBarStyles();
   const { t } = useTranslation();
+
+  //========================================================================================
+  /*                                                                                      *
+   *                                        Helpers                                       *
+   *                                                                                      */
+  //========================================================================================
+
+  /**
+   * Format option label
+   * @param {object} option : Node Data
+   * @returns {string} Option label
+   */
+  const getOptionLabel = option => {
+    const template = option.Template || option.ContainerFlow;
+    return `${option.name} [${template}]`;
+  };
 
   //========================================================================================
   /*                                                                                      *
@@ -97,14 +112,12 @@ const FlowSearch = props => {
   return (
     <Autocomplete
       data-testid="input_search-flow-text"
-      options={options}
-      getOptionLabel={option => option.name}
+      options={options.sort((a, b) => a.parent.localeCompare(b.parent))}
+      getOptionLabel={getOptionLabel}
       onChange={handleSearchNode}
       onBlur={handleSearchToggle}
       onFocus={onEnabled}
-      groupBy={option =>
-        option instanceof NodeInstance ? t("Node") : t("SubFlow")
-      }
+      groupBy={option => option.parent}
       PopperComponent={renderSearchPopup}
       renderInput={renderSearchInput}
     />
@@ -113,7 +126,7 @@ const FlowSearch = props => {
 
 FlowSearch.propTypes = {
   visible: PropTypes.bool,
-  options: PropTypes.arrayOf(PropTypes.string),
+  options: PropTypes.arrayOf(PropTypes.object),
   onChange: PropTypes.func,
   onEnabled: PropTypes.func,
   onDisabled: PropTypes.func