+++ src/editors/Flow/view/Components/Nodes/BaseNode/TreeView/TreeNode.js
@@ -1,15 +1,16 @@
 import * as d3 from "d3";
 import _debounce from "lodash/debounce";
+import { belongLineBuilder } from "../../Utils";
 import BaseNode from "../BaseNode";
+import BaseNodeStatus from "../BaseNodeStatus";
 import TreeNodeHeader from "./TreeNodeHeader";
 import TreeNodePort from "./TreeNodePort";
 import CollapsableItem from "./CollapsableItem";
-import { belongLineBuilder } from "../../Utils";
-import BaseNodeStatus from "../BaseNodeStatus";
 
 class TreeNode extends BaseNode {
   constructor({ canvas, node, events, _type, template, parent }) {
     super({ canvas, node, events, _type, template });
+    this.isContainer = false;
     this.parent = parent;
     this.children = [];
     this._links = new Map();
@@ -68,7 +69,8 @@ class TreeNode extends BaseNode {
       this.headerPos.y,
       this.name,
       this.templateName,
-      this._type
+      this._type,
+      this.data.endless
     );
 
     // append to the svg element
@@ -150,6 +152,7 @@ class TreeNode extends BaseNode {
       .renderStatus();
 
     if (addLinks) this.updateLinks();
+
     return this;
   };
 
@@ -171,6 +174,19 @@ class TreeNode extends BaseNode {
         y + nodeContainer.y.baseVal.value + parentContainer.y.baseVal.value;
     }
     return { x: nodePosX, y: nodePosY };
+  }
+
+  /**
+   * @override Get node center
+   * @returns {{xCenter: number, yCenter: number}}
+   */
+  getCenter() {
+    const { x, y } = this.getAbsolutePosition(this);
+
+    return {
+      xCenter: x + this.width / 2,
+      yCenter: y + this.height / 2
+    };
   }
 
   /**
@@ -384,10 +400,38 @@ class TreeNode extends BaseNode {
    *
    * @param {string} templateName node's template name
    */
-  onTemplateUpdate = templateName => {
+  onTemplateUpdate = template => {
+    const templateName = template.Label ?? template;
     if (templateName !== this.templateName) return; //not my template
     this._template = undefined;
+
+    if (this.isContainer) {
+      // Remove old belong line
+      this.removeBelongLine();
+      this.removeAllChildren();
+      this.removeCollapsibleElement();
+
+      // Re-add the subflow
+      this.canvas.mInterface.graph.updateSubFlow(this, template);
+    }
+
     this.update(true);
+  };
+
+  /**
+   * Removes all children and clears the array
+   */
+  removeAllChildren = () => {
+    this.children.forEach(child => child.destroy());
+    this.children = [];
+  };
+
+  /**
+   * Removes the NODES collapsible item
+   */
+  removeCollapsibleElement = () => {
+    this.collapsableItem.destroy();
+    this._collapsableItem = null;
   };
 
   /**