+++ CHANGELOG.md
@@ -19,7 +19,10 @@
 - [FP-1839 - New snackbar is crashing in all apps](https://movai.atlassian.net/browse/FP-1839)
 - [FP-1853 - Migrate Flow shortcuts - Reset Zoom and move nodes with arrow keys](https://movai.atlassian.net/browse/FP-1853)
 - [FP-1913 - Add editing capabilities to Tree View Editor](https://movai.atlassian.net/browse/FP-1913)
-- [FP-1913 - Added Exposed Ports validation and fix to remove them](https://movai.atlassian.net/browse/FP-1913)
+- [FP-1931 - Added Exposed Ports validation and fix to remove them](https://movai.atlassian.net/browse/FP-1931)
+- [FP-1993 - Tree View subflows now correctly update changes made to them](https://movai.atlassian.net/browse/FP-1993)
+- [FP-2007 - Avoid recursive template rendering](https://movai.atlassian.net/browse/FP-2007)
+- [FP-2008 - Composed the names of the subflow children on Tree View](https://movai.atlassian.net/browse/FP-2008)
 
 # 3.0.0
 
@@ -38,6 +41,7 @@
 - [FP-1595 - Removed unused function that contained vulnerabilities](https://movai.atlassian.net/browse/FP-1595)
 - [FP-1610 - Remove editor keybinds when editor is blurred](https://movai.atlassian.net/browse/FP-1610)
 - [FP-1838 - Give focus to active editor if it becomes dirty](https://movai.atlassian.net/browse/FP-1838)
+- [FP-1978 - Add Description field when Adding/Editing parameter on Flow Template](https://movai.atlassian.net/browse/FP-1978)
 
 **Callback Editor**